moved to cmd/collector/app/server/httpmetrics
